Added an array storage class with copy-on-write.
Added a new array storage class called Array, which has copy-on-write
semantics. This offers us the ability to return by value, and only make
deep copies of arrays if the array is modified while retaining the
ability to bypass most of this if using references or pointers.

This version of the templated class is not thread-safe. Using an atomic
int for the reference count, and a mutex during the copy operation will
allow us to make this container safe to share across threads while
retaining the copy-on-write capability.

Change-Id: I177882b845aa0001581b2e0774779924d101ac93