Define GLEW_STATIC on Windows
We are using the built-with-CMake GLEW on Windows.
For now, just define GLEW_STATIC because we know
the SuperBuild's glew32.lib is going to be static.

A better approach would be to build both static and
dynamic versions, and then explicitly link to the static
one. With this approach (which will work for now) we
run the risk of linker errors later on if somebody
switches the GLEW build to BUILD_SHARED_LIBS but does
not also change this add_definitions call here...

Change-Id: I233e60b160232b6bea2f81b0a4211607bb381bdd