Add very basic editing functionality.
This patch adds as much editing as possible with the current Molecule
container:

* Click in free space to add a new atom. Dragging will create another
  atom of the same type bonded to the first.
* Click an existing atom to change its type.
* Click and drag on an existing atom to add a new bonded atom.
* Click a bond to change its bond order.

The old Editor class has been renamed Manipulator, and can be used to
modify the positions of existing atoms and change bond orders. Event
handling has changed to send events to a single active tool, with
unhandled events going to the Navigator class.

Change-Id: I02264fefd48d70782d920d055e3495d32b1c4602