Adapt the VTK image diffing to work with AvogadroLibs
Create a Qt derived GLWidget, show it, render and capture the image.
Then duplicate the logic to compare with a baseline image (as used in
VTK) and produce a gamma shifted diff on failure. Use the Dart tags to
specify the results and upload the images upon failure.

Still to do - handle multiple baseline images.

Change-Id: I73bd085df7cbafa1c2a7cb858a7f2b75ee5ab00c