Refactor the classes for quantum mechanics
Split the core quantum classes into several, using Core::BasisSet as the
core data container, with Core::GaussianSet and Core::SlaterSet as the
derived types. Added a Core::GaussianSetTools to provide tools to
calculate properties such as molecular orbital values. Moved the
QtConcurrent based code into the quantum plugin to perform parallel
calculations of cubes (in much the same way as before) for molecular
orbitals and electron densities.

Ported the GAMESS-US, Gaussian and Molden readers to use the
Io::FileFormat base class, and removed the Qt dependecy from them. The
others will follow, using the same deign. This removes the need for the
BasisSetLoader class, and adds several enhancements to the GAMESS and
Gaussian parsers. They both need more work (and some example files) in
order to handle open shell calculations correctly.

This currently reads in the SlaterSet from the MOPAC AUX file, but the
cube calculation duplicates the GaussianSetTools class currently. The
proted code appears to work, but there is a need to validate the
calculated results against known base lines (this was true for the
previous versions of the code too).

Change-Id: Id05e60fef70e1af18f931880c2322f33e0049187