Add MoleQueueManager, MoleQueueDialog, MoleQueueWidget.
The manager class provides consistent access to a MoleQueue client
instance. It adds a layer on top of the MQ client that
maintains queue/program state in an MVC model (queueListModel())
and keeps a single client connected and available for use by the
application.

The InputGeneratorWidget is split to create a MoleQueueDialog
and MoleQueueWidget, which implements a reusable job submission
UI.

Other misc fixes/changes in this patch:
* More verbose error message in inputgenerator.cpp
* Set window title in InputGeneratorDialog
* Remove the 'settings' object in --generate-input stdin JSON. These
  should now be regular options. Updated existing generator scripts
  and GAMESS input plugin.

The client-value-api MoleQueue branch is needed for this to compile.

Change-Id: I41e695e187986c44d521b6839053855581ff623c