Cleaned up the dependencies and option logic
Removed an link to AvogadroRendering from AvogadroQtGui that was not
necessary, removed links to libraries already in the interface libraries
for dependent libraries. Removed the USE_BOOST option as this is a
simple fallback if C++11 is not available, and added a USE_VTK option.
Updated the testing directory to respect the options set.

Change-Id: I7ad6649184df601a0393d0d455d98905467352b2