Added a simple wrapper for mutex
This allows us to use mutex in the core API, enabling the migration of
Cube and Mesh classes into core without the loss of simple threading
constructs.

Note that try_lock requires us to link to pthread when using the Boost
fallback, added logic to the CMake code in STL to make that happen.

Change-Id: Id2681529ee24501f879b15a958cdbed4e45ee82c