WIP: Text rendering.
Still need better positioning, but this branch currently will
rendering a string at the origin in the xy plane.

See temporary modifications in glwidget.cpp for an example usage
of the text API.

This breaks other geometry rendering at the moment because of
changes in ShaderProgram attribute strides. This will be fixed soon.

Change-Id: Ic6186b76477760e60afd7cd9a404b6453670ae1f