Update text rendering to be easier to use.
This replaces the TextLabel + QuadPlacementStrategy approach with
two classes, TextLabel2D and TextLabel3D. Both are billboarded and
will render at the same size regardless of camera/label position.
The 2D label is anchored in window coordinates, while the 3D version
is anchored to a point in the scene.

Additionally, Texture2D is no longer a Drawable, but rather an
abstraction of the OpenGL API that presents texture information
in a class interface, much like BufferObject.

A documentation fix for the Overlay2D origin is included.

Change-Id: Ic01210e10aa2d4c148d4ab386959af0e6600e42b