Add basic crystallographic support.
This adds a simple UnitCell class, adds a hook into Core::Molecule,
and creates a CrystalTools algorithm collection, which is a port of
various crystallographic algorithms from Avogadro 1's crystallography
extension and XtalOpt.

The Crystal extension exposes much of the CrystalTools functionality
to the user, and the CrystalScene scene plugin provides simple
unit cell rendering.

Change-Id: Ib7c945cbed452287ce5ad77114529a6723462656