Port the QTAIM extension to Avogadro 2.
* Move everything into the Avogadro::QtPlugins namespace
* Remove the unfinished settings widget
* Update Molecule, Atom, and Bond usage
* Rewrite extension, sceneplugin/engine interfaces
* Update Eigen include in qtaimmathutilities.h
* Fix compile error using Eigen::Map with const input data
* Relicense to BSD where possible.

The qtaimcubature files appear to contain a copy/paste of
abinit's cubature library, which is GPL licensed and has roots
in older GPL projects. Thus the entire extension is left as a
GPL extension to the BSD licensed Avogadro until the cubature
implementation can be replaced/rewritten/relicensed.

A BUILD_GPL_PLUGINS cmake option is added to address this.

Change-Id: I66f093a444be99ea7a8395eea2878414744aa967