Added persistent atom and bond objects
This offers an object that can contain a persistent atom, reporting
correct validity based on whether the contained unique ID is actually
pointing to a an atom in the molecule or not. The same is true for
the persistent bond object, with both new objects tested in the
molecule tests.

Change-Id: Ia9fae2c2f6eca4e1bb987bb7603b2e37a621bae6