Added support for generating a movie file/GIF
This adds really rough support for generating images for a movie,
then calling the relevant command line tools to assemble the movie.
This still needs work t omake the size of the movie configurable, as
well as the location of the executables and their arguments.

Pushing this in as a work in progress, it may fail in not so obvious
ways as it also needs lots of extra error handling code too!

Change-Id: Ib2d4d4668b2356d4e3b049a22ed6bd9a064fa92e