Made QtGui::RWMolecule a QtGui::Molecule container
Operating on the Molecule data directly, rather than duplicating,
gives us most of the desired functionality while ridding us of the
need to convert back and forth between molecule types. This may
be refined further in the future, but in the short term it lets us
retain undo/redo while offering a simpler edit/view switch.