Fixed issue where progress dialog showed
This was seen on Arch Linux with Qt 5.5.0, and looks like it might be
a bug but on the other hand it is pretty wasteful for a plugin to
initialize an object that may never be used. We should always prefer
delayed initialization, and this seems to fix the issue for me.