Added a new spectra plugin - vibrational modes etc
This plugin generates the frames, and animates vibrational modes.
Some of these things might be better placed in the library so that
they can be reused in the future but this is a good place to build
them out for now.

You can view the vibrational modes, select one, animate it, change
the scale factor used, although the intensity is missing at this
point. That should get added in next.