fix Gaussian FCHK parsing of unrestricted calculations
Gaussian FCHK files have a propensity to not emit a UHF or RHF tag which is the only way the parser would check for the proper calculation type. This PR fixes this by double checking the calculation type if it finds a "Beta Orbital Energies" entry in the FCHK file and fixing the entries appropriately.