Provide some minimal Python wrapping using Boost Python
You know, I looked at so many other solutions, and I kept coming back
to this one. We do not need to wrap everything, and it offers a nice
mechanism for making things more Pythonic inline, works well with
Python 2 or 3, and we used it before.

Currently very simple, wrapping a few of the major classes for some
initial testing. I think it should be extended to expose each library
as a component, i.e. avogadro2.core, avogadro2.io, etc. The singleton
for file formats has a nice Pythonic conversion, Python strings are
immutable, and so the writeString method had to be manually wrapped.