First pass at writing out a cube in the CJSON
This needs work as it will only write out the first cube, and does not
describe what the cube is, i.e. molecular orbital 21, etc. It does
however give a starting point, and a method of using the API to
calculate the cube for an MO and write it out to CJSON.