Added POSCAR format and clipboard import crystal.
Added a POSCAR format reader and writer to the list of formats in Avogadro2.
Also added an "import crystal from clipboard" feature similar to that of
Avogadro1. It attempts to read the clipboard text using POSCAR format by
default. If a different format is to be used, the user can input that extension
and the crystal will be read for that format using obabel instead.