Moved hydrogentools test to qtgui and updated it
This fixes a compiling error when we compile the tests. HydrogenTools was
recently moved from core to qtgui, so we should move the test as well. In
addition, Molecule was changed to RWMolecule and Atom to RWAtom to reflect
the changes in the in the hydrogentools functions (because they now take
RWMolecule instead of Molecule). The hydrogen tools test passes when it is ran.