Added reduce to asymmetric unit function
A reduction to an asymmmetric unit basically means that
only one atom from each Wyckoff position will be kept - the rest
will be removed. Essentially, this takes the transforms of the atoms
checks to see if there is an atom lying on any of the transform
positions. If there is an atom there, it removes that atom.
It is essentially the opposite of filling a unit cell.

It seems to work fine. I added tests for it as well.