Fixed some bugs in the atomic coordinate editor
The atomic editor wasn't working at all for me originally because
movePosition(QTextCursor::NextCharacter) was not moving it to the
end of the line. movePosition(QTextCursor::Down) followed by
movePosition(QTextCursor::StartOfLine) seems to work fine! And I
have it checking to make sure it only moves if it is not at the
end of the document. Additional bug fixes: program no longer freezes
if the user opens up the coordinate editor and hits "enter" or
"return" on a blank document. In addition, the input validator skips
over empty lines - this is helpful if the user deletes some lines but
forgets to delete white space or returns. Finally, I also set it up
so that the undo stack is updated when the coordinates are edited,
and it is now "undo-able."