Added adjustHydrogens to RWMolecule
Added adjustHydrogens to RWMolecule as an undo command.
Also added adjustHydrogens for multiple atoms as a single
undo command. In addition, added a new addAtom() that combines
adding the atoms and setting its position into one command. This
reduces the number of undo commands that appear while drawing. More
still needs to be done to reduce the number of undo commads further,
but I believe this is a good start.