Adjust hydrogens after releasing left click
The code is currently set up to adjust hydrogens
upon left clicking an atom. If we drag away from the
atom, and then release, it has to adjust hydrogens
again. So I think it is better to just adjust hydrogens
upon releasing left click rather than on left click. This
is how it was in Avogadro1.