Fixed a bug when redo-ing Add Atom with no atoms
Before this fix: if a user had no atoms and tried to perform
a redo, addAtom() in RWMolecule would not add the atom correctly.
I added a boolean to AddAtomCommand to check and see if we are
using positions for the atom or not; it is set to true by default.
If we are not using positions (perhaps with something like SMILES),
it may be set to false as a second parameter for addAtom().

This boolean is basically what the "if (!positions.empty())"
statement was checking for - only it does it better because
the positions are unfortunately empty if there are presently
no atoms in the molecule.

There may be a better way to fix this, but this does work.