Fix remaining issues with the CJSON reader
The fractional coordinates need to have the unit cell in order to be used.
The readProperties should not be required, and should be renamed to reflect
the quantum data being read in. The reader shouldn't perceive bonds,
possibly optional based on whether there are bonds in the molecule or not.