Initial version of a selection tool
This only supports atom selection at this stage, with some support for
rendering a selection box that is disabled until the support is added to
the scene to determine what primitives are in the box area. This just
toggles atoms, and saves them to a selected atom array. Some basic
support was added to ball and stick to render selected atoms as blue and
a little bigger.

This is intended as a preview, but is far from complete - sorry it took
so long, hopefully I will have a more complete version by year end!