Added some error checking, commented out leaky code
Added a bit of error checking on property value code to ensure they are
seen as the type intended to be used. Commented out a significant amount
of code, keeping it for now so that we can decide what should be
integrated. There are two reasons I am commenting them out - it is
impossible to get back to the types stored in the variant, and as such
there is no way to integrate that data until new code is added.

The variant class should be improved to throw compile errors, or handle
some of these types. I think for all C-style arrays we should throw a
compile error as there is no way to determine the length of the array,
or properly manage the memory and C++ containers should be preferred.
For the Array class we could add code for some of the common types, but
need to consider whether it is worthwhile or it would be better to add
containers in the Molecule class (or supporting classes such as basis
set, unit cell, etc).