Fixed a few errors in spacegroupdata.h
As was pointed out in issue number 153 on
https://github.com/openchemistry/avogadrolibs, there
are a few issues in the space groups. There were a few
illegal '#' symbols and potential problems with Hall numbers
506 (international number 210) and 450 (international number
160). Unfortunately, I couldn't fix 450, but I was able to
fix 506. A test was performed: after using the transforms
in Avogadro, spglib identified the space group correctly.
For some reason, this was not the case for number 450...
These should be the right transforms here:
http://www.cryst.ehu.es/cgi-bin/cryst/programs/nph-getgen?what=text&gnum=160
but I haven't been able to get spglib to identify it correctly.