Fixed a compiling issue with array.h on Windows
For some reason, I was experiencing the following error
for some template types for Array:

"formal parameter with requested alignment of 16 won't be aligned"

and the error occurred on the altered line.

I don't fully know why, but some sources online suggested to use a
reference to fix this issue. So I changed the function to pass the
type by a const reference, and it seems to fix it. It looks like it
might have something to do with it not working on 32-bit systems.

As far as I can tell, it seems like a reasonable fix. Let me know
if you think otherwise, though. This fixed the problems.