Move the Python wrapping to PyBind11
I think PyBind11 is superior at this stage, and as it is header only
it makes building our (currently limited) Python wrapping simpler. As
we expand the API coverage we should split into separate Python
modules, but I think this is fine right now.

The support for singleton classes isn't quite as storng just yet, so I
added a simple proxy class to expose the file format manager with the
same syntax in a lightweight proxy class.