Add name property to basis set
The name of a basis set is useful if it is a simple basis set applied to
all atoms in a calculation. Add a name property, and set it to the name
of the uniformly applied basis set, or to custom if more than one basis
set is observed in a given file.