Added theoretical XRD pattern calculations
Added the ability to calculate and plot a theoretical XRD
pattern given a crystal.

This uses an external, static "genXrdPattern" executable to perform
the calculation. The source for this program can be found here:
https://github.com/psavery/genXrdPattern

The majority of the work comes from ObjCryst++, though:
https://github.com/vincefn/objcryst

VTK is used to generate the plot. As such, this plugin will only
be compiled if USE_VTK=ON.