Use the Qt support classes for the chart
This switches to use the QVTKOpenGLWidget, at which point the chart is
contained in a widget, using the same event loop, etc. This is a little
hackish, I think the plugin should lazily initialize the plot object on
first show, and then update the data/call show on subsequent clicks.

I think this at least shows the basics, and gives us a simple path to
explore other functionality.

Change-Id: Ibc27a4cef7b6fc802a2f0ca78a1ac1476a475f37