Improve the libmsym find calls
They should both just use lower case, and find should only be called
from the place where it is being used - the symmetry plugin. This logic
now finds the libmsym CMake config for the superbuild, and if not found
will use the CMake find package to look for a system copy.