Remove the old reader code, mostly replaced
The CJSON format had drifted, and gained some asymmetry on the read
and write side. The reader now reads almost all that it used to, and
will read what the writer puts in the file more consistently now.

This is pretty much entirely converted to the new JSON parser for
reading, the things that were missed can be found in history and
restored if desired (the CJSON test passes). The writer is next.

Signed-off-by: Marcus D. Hanwell <marcus.hanwell@kitware.com>