Added an active objects class to Avogadro
This enables us to avoid calling functions for every plugin, they can
query for active objects. Currently just one object as a proof of
concept, this can be expanded out. This lets us keep the API reasonably
clean, and have a singleton manage access to some of the useful global
objects within a running application.

Signed-off-by: Marcus D. Hanwell <marcus.hanwell@kitware.com>