Add ability to animate through orbitals
This augments the dialog to make it possible to animate through a
sequence of electronic structure calculations. It is still at an early
stage, but is quite robust and makes a number of improvements and bug
fixes.

The mesh and cube management is vastly improved, locking, and triggering
are all more robust. Switching between detail levels, or between
molecules, or between frames is all far more robust. Reuse some of the
player tool code, but coordinate an asynchronous rendering of frames.

Signed-off-by: Marcus D. Hanwell <marcus.hanwell@kitware.com>