Bring VTK volume rendering and surfaces to Avogadro
This is a pretty large commit, I had the intention of breaking it up
more, but it is all highly related and hard to bring in piece-by-piece.
There are a few things to note, such as the fact that VTK now requires
an OpenGL 3.3 core profile, making it incompatible with the Avogadro
rendering. This means that we must defer all rendering to VTK in the VTK
render window, and copy our data into VTK data structures.

The vtkImageData format also uses a different ordering (often referred
to as "Fortran ordering", whereas our Cube class uses the "C ordering"
used in most places (C, C++, Python, etc). So cubes must be reordered
as they are copied over.

The color opacity widget is ported from Tomviz, with some adaptations to
only use VTK, and some work to create a color map that looks good for
typical molecular orbitals. This is needed to see the populations of
values in cubes, and modify the color opacity map for volume rendering.

The widget and the view both update as new cubes are calculated, and
trigger new renders. There is still some work to do, and the active
objects work needs the corresponding additions in the application to
ensure we can see what is currently active. It is very cool to see MOs
pop in, and to benefit from the speed of the flying edges algorithm.
This also reuses cubes, which the standard approach should do too.

Signed-off-by: Marcus D. Hanwell <marcus.hanwell@kitware.com>