Core Molecule: Cache the most recent space group
This can save time if a plugin wishes to attempt to use a cached
space group rather than determine it algorithmically via spglib every
time it is needed.

This can also enable use to put space group information in the corner
render window of a crystal if it is desirable.

All of the information about a space group (the crystal system,
international number, schoenflies symbol, hall symbol, international
symbol, etc.) can be determined with just the hall number
(functions to do these are in avogadro/core/spacegroups.h).

Signed-off-by: Patrick Avery <psavery@buffalo.edu>