Yaehmop band structure: Generate special k points
This commit enables special k points to be generated automatically
using the space group of the crystal. It will first attempt to use
the cached space group, and if that does not exist, it will attempt
to use spglib to determine the space group.

It seems to work in all of the standard test cases that I tried.

Signed-off-by: Patrick Avery <psavery@buffalo.edu>