AppVeyor: Use ninja for the build
Ninja seems to be much faster for the build than the
previous build system, partially because it uses all
of the cores on AppVeyor.

Ninja is installed with chocolatey, and the Visual Studio 2017
environment variables are set with the `vcvarsall.bat` script.

Verbosity detail was also turned off. This may have caused a
significant speed increase as well.

Signed-off-by: Patrick Avery <psavery@buffalo.edu>