Added ability to search for MongoChem molecules
A user can now search for molecules on a MongoChem server via
"Extension"->"MongoChem". The "Search" button populates the table with
entries where the rows are different molecules and the columns are
"Formula", "SMILES", and "InChIKey".

This is a simple draft setup where there are several hard-coded values
(including the web address of the mongo chem server as
"localhost:8080"), and a QTableWidget is used instead of a model/view
approach. The user also cannot log in, so it just shows public
molecules.

More will be added in the future...

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>