Added config dialog
In the config dialog, a user can set the base URL for the requests,
and they can also set their api key. The api key line edit is in
password mode (so that the characters appear as dots).

Authentication is not yet performed with the api key, so it is
currently not being used for anything. But authentication will
be added soon.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>