Added authentication after config is accepted
If the config is accepted, and there is an api key, authentication is
attempted. Currently, a message box appears indicating whether the
authentication succeeded or failed.

Once authenticated, the girder token will be used for future requests.

To do these things, a new "post()" method was added to the
GirderRequest class.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>