Added ability to download molecules
A "Download" button was added at the bottom of the widget so
that if a row is selected, the molecule in that row may be
downloaded and set in Avogadro2.

If the molecule has a name, the name will be used in Avogadro2
as well. If it does not have a name, the formula will be used.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>