Added support for uploading molecules
If a user is logged in and a molecule is drawn in the Avogadro2 window,
the molecule can be uploaded to mongochemserver.

Note, however, that geometries are not currently supported. This means
that if a user tries to upload a molecule that has the same inchikey
as a molecule in the database, a new molecule will not be created
(additionally, it still reports "Upload succeeded" when this happens
because there is currently no way to know from the post response whether
a new molecule was created or not).

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>