Added the MoleQueue client code to Avogadro
This removes the optional dependency on the MoleQueue client library.
This was always a weak point of the MoleQueue design, and I was never
happy with the hard dependency on MoleQueue for a tiny client library.
It just uses Qt's Network module with some JSON for coordinating job
submission. Build this unconditionally when Qt is enabled, leave the
directory naming alone.

Signed-off-by: Marcus D. Hanwell <marcus@cryos.net>