GAMESS-US output reader sanity check
The GAMESS-US output reader now raises an error if it could not read any atomic coordinates from the output. Also make the return value nodiscard to force callers to check if the read actually succeeded, as this function also alters the state of a Molecule object if the read succeeds.

Signed-off-by: GYT <tiborgyri@gmail.com>
Signed-off-by: Tibor Győri <tiborgyri@gmail.com>