Fix inconsistent type in the cube creation code
I have to be honest I am not sure how that ever worked! This code
reorders the cube from the Avogadro ordering to the VTK ordering, but it
was incorrectly casting to double* rather than float*. This was causing
a segfault in my testing, once fixed it works as expected.

Signed-off-by: Marcus D. Hanwell <marcus@cryos.net>