Ensure the VTK plot widget has a minimum size
This is very hackish, but it allows you to see the plot. The main issue
is that this should have always been in a dialog or similar with a
layout. The way that VTK interacts with Qt has changed and so the size
being set does not appear to act as it once did. Issue #1107.

Signed-off-by: Marcus D. Hanwell <marcus@cryos.net>