Adding support for Qt 6 with 5 still an option
Qt 5 is end of life, this is my first pass at porting to Qt 6. This
commit just contains the CMake changes so that is is easier to review
these separately from all the code changes in a follow up commit.

Note that I have put a number of plugins behind a variable guard and
they only build on Qt 5 for now. This will allow you to continue using
everything with Qt 5 while developers can port their code to Qt 6 and
test it at will. I took a pass at the simpler things in this pass.

Signed-off-by: Marcus D. Hanwell <marcus@cryos.net>