Remove avogadrogl.h from the SolidPipeline header
The basic design of the rendering code is to only include OpenGL in the
implementation files, not in the public headers. This commit isn't
looking at what the class is doing, just moving all OpenGL/GLEW code to
the implementation, and using a simple d pointer. This header is
included by glrenderer.h and included in several places.

I didn't follow the addition of this code and the layers, I verified
things look the same before and after these changes. The inclusion of
any OpenGL/GLEW headers in the API is generally problematic and should
be avoided. Keeping these things in the implementation offers a level of
encapsulation with little to no performance impact.

Signed-off-by: Marcus D. Hanwell <marcus@cryos.net>