Remove the export include in avogadrocore.h
This commit removes the include of the export header in avogadrocore.h,
which is mainly there to define common symbols and was only adding it as
a convenience. Took a first pass at adding the export macro include to
the headers that need it, and where obvious removing the avogadrocore.h
include if it was not being used.

Signed-off-by: Marcus D. Hanwell <marcus@cryos.net>