Move to a more modern CMake build infrastructure
There is more to come, but this is a large leap forward. It embraces the
newer target properties for source files, header files (requiring the
bump to CMake 3.23), and separates out some header only components in
core that were used without linking. There is an argument to moving
these to a separate directory but I kept it simple for now.

The avogadro_headers function wraps some boiler plate for adding headers
to a target that will get installed. The targets use the base name, and
an alias is added with the namespace to expose them in the same way as
outside users would see them. The use of namespaces is disruptive but
worth it as CMake will always interpret this as a target so that errors
are raised at configure time rather than link time.

There is still some legacy stuff, but got rid of a few global
include_directories calls that revealed other issues. In general we had
limited include_directories to directory scopes, but this is much more
granular. Generator expressions let us easily express the build and
installed include directories. A follow up commit will migrate all of
the plugins.

Signed-off-by: Marcus D. Hanwell <marcus@cryos.net>