Porting the plugins to use the new target names
This was some liberal use of sed along with testing everything linked.
There may be one or two not currently being built that were caught by
sed, there is still a fair bit of calls that should be removed when
things like Eigen are carried around by the target properties.

Signed-off-by: Marcus D. Hanwell <marcus@cryos.net>