Modernize Eigen3 use as an imported target
Eigen3 is a header only template library, adopt an updated
FindEigen.cmake that creates an imported target. Note that include
directories from imported targets are automatically treated as system
includes on compilers that support the concept.

The core header only Avogadro library exposes it as part of its
interface, and then we are able to remove a huge amount of duplicated
code. Still need to look at the other side to ensure the exported
targets find all targets like Eigen they depend upon.

Signed-off-by: Marcus D. Hanwell <marcus@cryos.net>