Include files need to be a little more convoluted
In the (fairly common) situation where someone builds in a build
subdirectory of the source tree the FILE_SETS command would complain as
the PROJECT_BINARY_DIR is a subdirectory of the PROJECT_SOURCE_DIR and
hence is not allowed. This change is a compromise to install to the same
location, but to next a little more deeply to allow build directories as
subdirectories of the source tree.

It has the side effect of adding extra directories to the include
directories in the build tree, there does not appear to be a way around
that. This works and supports a pretty common usage pattern I think we
should support. I discovered this while looking at the scikit build for
the Python wheels.

Signed-off-by: Marcus D. Hanwell <marcus@cryos.net>