add UniqueID label type
If UniqueID label type is chosen, atoms are numbered from zero,
in contrast to index, where atoms are human-numbered from unity.

UniqueID mode is needed e.g. for chemists working with Orca quantum chemistry software. It is common to specify atoms in Orca input file (e.g., for a forming bond in a TS search), and Orca counts from zero.

Signed-off-by: Ivan Mitrichev <imitrichev@users.noreply.github.com>